pyproject.toml
[project]
name = "bookwriter-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.82.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
    "yaspin>=3.1.0",
]
